Added mocks file
