+++ src/components/dish/dish.hbs
@@ -1,11 +1,11 @@
-<div href="dish/{{id}}" class="dish">
+<div href="dish/{{id}}" id="{{id}}" class="dish">
     <div class="dish__text">
         <div class="dish__title">
             {{dishTitle}}
         </div>
 
         <div class="dish__info">
-            {{dishCost}}  • {{dishCcal}}
+            {{dishCost}} ₽  • {{dishCcal}}
         </div>
     </div>
     <img class="dish__image" src="{{dishImg}}">
