+++ src/views/restaurantView/restaurantPage.hbs
@@ -1,5 +1,5 @@
  <div class="restaurant-page__dishes">
-     {{#each menuBlock}}
+     {{#each menuNavs}}
          <div id="{{menuId}}" class="restaurant-page__menu-title">
              {{menuTitle}}
          </div>
